Fixed bugs in makefile
Bug 1:

For some users, the build dir is not automatically created. It is also required for javac "The directory must already exist because `javac` does not create it." See `man javac` under the -d flag

Fixed by adding `mkdir build` to the NodeChan.jar section.

Bug 2:
Since `make jar` requires that NodeChan.jar to be created, make does it for us without explicitly being told to, which made it run twice.

Fixed by removing the redundant call to `make` in the jar section.